Update 01d_two_brightnesses.py
I changed the list to count up from 0.2 to 1.0 and back down to 0.2. I also added a 0 at the end of that list so that the light would turn off for a second after it snores. Finally I changed the number repeats to 30 so that it snores quicker than it would if that number was 100.